requirements.txt
transformers==4.40.0
torch
torchvision
timm
Pillow
pyttsx3
TTS
